Use Nettle's arcfour rather than carrying a local implementation
This also reduces the default seed length to 256 bytes. The arcfour
RNG would only use the first 256 bytes anyway and the others use much
less than that so there is no functional change.

In addition there is now no FSF code in the source, so the copyright
notices are updated accordingly.  (But note that object files attract
an interest from many copyright holders due to the use of Nettle.)